test(doc): added tests on metadata keys with glob patters (*, **)
